LPS-195773 Frontend Tests
